feat: add validation in section mutations

+++ src/sections/mutations/deleteSection.ts
@@ -1,18 +1,27 @@
-import { resolver } from "@blitzjs/rpc";
-import db from "db";
-import { z } from "zod";
+import { resolver } from "@blitzjs/rpc"
+import db from "db"
+import getChapter from "src/chapters/queries/getChapter"
+import { z } from "zod"
+import getSection from "../queries/getSection"
+import getBook from "src/books/queries/getBook"
 
 const DeleteSection = z.object({
   id: z.string(),
-});
+})
 
 export default resolver.pipe(
   resolver.zod(DeleteSection),
   resolver.authorize(),
-  async ({ id }) => {
-    // TODO: in multi-tenant app, you must add validation to ensure correct tenant
-    const section = await db.section.deleteMany({ where: { id } });
+  async ({ id }, ctx) => {
+    const currentUserId = ctx.session.userId
+    const section = await getSection({ id }, ctx)
+    const chapter = await getChapter({ id: section.chapterId }, ctx)
+    const book = await getBook({ id: chapter.bookId }, ctx)
 
-    return section;
+    if (currentUserId != book.userId) return
+
+    await db.section.deleteMany({ where: { id } })
+
+    return section
   }
-);
+)
